requirements.txt
# Telegram Bot API
python-telegram-bot>=20.3

# PostgreSQL async driver
asyncpg>=0.28.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client (для будущих интеграций)
aiohttp>=3.8.0

# Timezone support
pytz>=2023.3

# Дополнительные утилиты
